pyproject.toml
[project]
name = "spy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "google-adk>=0.2.0",
    "litellm==1.66.1",
    "mcp[cli]>=1.6.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]
